+++ src/validators/kropki-validator/validateKropki.ts
